pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "adafruit-circuitpython-adafruitio>=6.0.3",
    "flask>=3.1.1",
    "requests>=2.32.4",
]
